IcPartSMASH now allows for >1 initial state SMASH
In such case, contributions from the input hadrons are scaled by
1/nevents. This way one can produce an averaged dynamical initial state.

To allow for that, hadrons from all initial state events are read into a
common vector, then time-ordering is performed, then all hadrons are
pushed into the existing queue. The number of initial state events is
computed as the number of lines starting with "#" and ending with "start"
in the SMASH IC file.